Fix macro-redefined warnings caused by limits.h & float.h (#23522)
* Do not use ULONG_MAX in favor of UINT32_MAX
* Do not use LONG_MAX in favor of INT32_MAX

Fixed size 32-bit LONG_MAX conflicts with standard macros
in limits.h on LP64 systems.

* Do not use FLT_MAX, and DBL_MAX macros when STDCPP_COMPAT is defined
* Do not define LONG_MIN, LONG_MAX, ULONG_MAX in pal/inc/pal.h
* Remove duplicated *_MAX macros in pal/inc/rt/intsafe.h

Fixes #20784

+++ src/debug/daccess/daccess.cpp
@@ -6321,9 +6321,9 @@ bool ClrDataAccess::ReportMem(TADDR addr, TSIZE_T size, bool fExpectSuccess /*= 
     while (size)
     {
         ULONG32 enumSize;
-        if (size > ULONG_MAX)
+        if (size > UINT32_MAX)
         {
-            enumSize = ULONG_MAX;
+            enumSize = UINT32_MAX;
         }
         else
         {
